feat: changed UserProvider API
Moved withAuthGuard to UserProvider and changed the provider behaviour

BREAKING CHANGE: withAuthGuard no longer exists outside the UserProvider
User will always be saved
and loaded as a cookie
