--- src/hoc/index.ts
@@ -1,1 +1,0 @@
-export { default as withAuthGuard, AuthGuardConfig } from './withAuthGuard/withAuthGuard';
