+++ src/providers/createUserProvider/createUserProvider.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import React from 'react';
-import client from '../createHttpClient/createHttpClient';
+import client from '../../utils/createHttpClient/createHttpClient';
 import createUserProvider from './createUserProvider';
 import { UserProviderUrls, UserProviderFactory } from './types';
 
@@ -41,7 +41,7 @@ describe('createUserProvider', () => {
     postSpy.mockRestore();
   });
 
-  describe('fetch mode', () => {
+  describe('', () => {
     let factory: UserProviderFactory<User>;
     let Component: React.FC;
     const userData: User = { id: 1, email: 'test@test.com' };
@@ -50,12 +50,11 @@ describe('createUserProvider', () => {
       const mockFetch: jest.Mock<User> = jest.fn();
       factory = createUserProvider<User>({
         ...providerOptions,
-        mode: 'fetch',
         useProfile: mockFetch.mockImplementation(() => userData),
       });
     });
 
-    it('returns a provider', async () => {
+    it('should return a provider', async () => {
       Component = () => {
         const { user } = factory.useUser();
         return <div data-testid="user">{JSON.stringify(user)}</div>;
@@ -71,7 +70,7 @@ describe('createUserProvider', () => {
       expect(JSON.parse(node.textContent || '')).toMatchObject(userData);
     });
 
-    it('updates the user', async () => {
+    it('should update the user', async () => {
       const updatedUser = { id: 1, email: 'dummy@test.com' };
       Component = () => {
         const { user, update, logout } = factory.useUser();
@@ -101,71 +100,6 @@ describe('createUserProvider', () => {
       expect(JSON.parse(node.textContent || '')).toMatchObject(userData);
       fireEvent.click(loginBtn);
       expect(JSON.parse(node.textContent || '')).toMatchObject(updatedUser);
-      fireEvent.click(logoutBtn);
-      await waitFor(() => {
-        expect(JSON.parse(node.textContent || '')).toEqual(null);
-      });
-      expect(postSpy).toHaveBeenCalledWith(providerOptions.logoutUrl, undefined, { withCredentials: true });
-    });
-  });
-  describe('storage mode - cookie', () => {
-    let factory: UserProviderFactory<User>;
-    let Component: React.FC;
-
-    beforeEach(() => {
-      factory = createUserProvider<User>({
-        ...providerOptions,
-        mode: 'storage',
-        storage: 'cookie',
-        key: 'default',
-      });
-    });
-
-    it('returns a provider', () => {
-      Component = () => {
-        const { user } = factory.useUser();
-        return <div data-testid="user">{JSON.stringify(user)}</div>;
-      };
-
-      render(
-        <factory.UserProvider>
-          <Component></Component>
-        </factory.UserProvider>
-      );
-
-      const node = screen.getByTestId('user');
-      expect(node.textContent).toEqual('null');
-    });
-
-    it('updates the user', async () => {
-      Component = () => {
-        const { user, update, logout } = factory.useUser();
-        return (
-          <div>
-            <div data-testid="user">{JSON.stringify(user)}</div>
-            <button data-testid="login" onClick={() => update({ id: 1, email: 'dummy@test.com' })}>
-              Login
-            </button>
-            <button data-testid="logout" onClick={() => logout()}>
-              Logout
-            </button>
-          </div>
-        );
-      };
-
-      render(
-        <factory.UserProvider>
-          <Component></Component>
-        </factory.UserProvider>
-      );
-
-      const node = screen.getByTestId('user');
-      const loginBtn = screen.getByTestId('login');
-      const logoutBtn = screen.getByTestId('logout');
-
-      expect(node.textContent).toEqual('null');
-      fireEvent.click(loginBtn);
-      expect(JSON.parse(node.textContent || '')).toMatchObject({ id: 1, email: 'dummy@test.com' });
       fireEvent.click(logoutBtn);
       await waitFor(() => {
         expect(JSON.parse(node.textContent || '')).toEqual(null);
