+++ src/providers/index.ts
@@ -1,0 +1,9 @@
+export { default as createUserProvider } from './createUserProvider/createUserProvider';
+export {
+  AuthGuardConfig,
+  DefaultUser,
+  UserObject,
+  UserProviderConfig,
+  UserProviderFactory,
+  UserProviderUrls,
+} from './createUserProvider/types';
