+++ src/utils/index.ts
@@ -1,15 +1,5 @@
 export { default as cookies, createCookieName } from './cookies/cookies';
-export { default as createAuthGuard } from './createAuthGuard/createAuthGuard';
 export { default as httpClient, createHttpClient } from './createHttpClient/createHttpClient';
-export { default as createUserProvider } from './createUserProvider/createUserProvider';
-export {
-  DefaultUser,
-  UserObject,
-  UserProviderConfig,
-  UserProviderFactory,
-  UserProviderMode,
-  UserProviderUrls,
-} from './createUserProvider/types';
 export {
   default as login,
   isLocalCredentials,
