+++ src/utils/login/login.ts
@@ -1,26 +1,27 @@
 import axios, { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
 import IResponse from '../../interfaces/IResponse';
-import { DatabaseRecord, HttpClient } from '../../interfaces/Record';
+import { DatabaseRecord, HttpClient, Prettify } from '../../interfaces/Record';
 import cookies from '../cookies/cookies';
 
 type LocalProvider = 'local';
 type SocialProvider = 'google' | 'facebook';
-type SocialPayload = SocialCredentials & { social_provider: string };
+type SocialPayload = Prettify<SocialCredentials & { social_provider: string }>;
 type AuthResponse<TUser extends DatabaseRecord> = IResponse<{ [x: string]: TUser }>;
 
 export type AuthenticatedUser<TProps extends DatabaseRecord> = { token: string } & TProps;
 
-export type Credentials = LocalCredentials | SocialCredentials;
+export type Credentials = Prettify<LocalCredentials | SocialCredentials>;
 export type LocalCredentials = { email: string; password: string };
 export type SocialCredentials = { social_token: string };
 
 export type Provider = LocalProvider | SocialProvider;
 
-export type LoginConfig = { apiUrl: string } & LoginKeys & LoginProvider & HttpClient;
+export type LoginConfig = Prettify<{ apiUrl: string } & LoginKeys & LoginProvider & HttpClient>;
 export type LoginKeys = { dataKey: string; authKey?: string };
 export type LoginProvider =
   | { provider: LocalProvider; credentials: LocalCredentials }
   | { provider: SocialProvider; credentials: SocialCredentials };
+
 /**
  * Provides an easy way to get an access token for the `fotexnet` infrastructure.
  * Define the login endpoint via `apiUrl`, the `dataKey` which will be used to identify the user object
@@ -43,7 +44,8 @@ async function login<TRecord extends DatabaseRecord = DatabaseRecord>(
   const user = res.data.data[config.dataKey];
   const authKey = config.authKey || 'authorization';
   const token = res.headers[authKey]?.split(' ')[1] || null;
-  cookies.set(authKey, token, 365 * 1000);
+  cookies.set(config.dataKey, JSON.stringify(user), 365);
+  cookies.set(authKey, token, 365);
 
   return { ...user, token };
 }
